backend/requirement.txt
flask==3.0.3
mysql-connector-python==8.4.0
flask-cors==4.0.1
requests==2.32.2
paho-mqtt==2.1.0
getmac==0.9.4
